refactor: Moved field to new component
