+++ src/components/PiecesPreviewModal/PiecesPreviewModal.js
@@ -1,6 +1,5 @@
 import { useState } from 'react';
 import PropTypes from 'prop-types';
-import { Field } from 'react-final-form';
 import arrayMutators from 'final-form-arrays';
 import { useMutation } from 'react-query';
 import { useHistory } from 'react-router-dom';
@@ -8,37 +7,22 @@ import { FormattedMessage } from 'react-intl';
 
 import { useOkapiKy } from '@folio/stripes/core';
 import { FormModal } from '@k-int/stripes-kint-components';
-import {
-  Datepicker,
-  Row,
-  Col,
-  Button,
-  MultiColumnList,
-  Label,
-  TextField,
-  Layout,
-  TextArea,
-} from '@folio/stripes/components';
+import { Row, Col, Button, MultiColumnList } from '@folio/stripes/components';
 
-import {
-  composeValidators,
-  requiredValidator,
-} from '@folio/stripes-erm-components';
-
-import { urls, validateWithinRange } from '../utils';
-
-import css from './PiecesPreviewModal.css';
+import { urls } from '../utils';
 
 import {
   CREATE_PREDICTED_PIECES,
   GENERATE_PIECES_PREVIEW,
 } from '../../constants/endpoints';
 import PiecePublicationDate from '../PiecePublicationDate';
+import PiecesPreviewModalForm from './PiecesPreviewModalForm';
 
 const propTypes = {
   showModal: PropTypes.bool,
   setShowModal: PropTypes.func,
   ruleset: PropTypes.object,
+  pieceSets: PropTypes.arrayOf(PropTypes.object),
   allowCreation: PropTypes.bool,
 };
 
@@ -46,6 +30,7 @@ const PiecesPreviewModal = ({
   showModal,
   setShowModal,
   ruleset,
+  pieceSets,
   allowCreation = false,
 }) => {
   const ky = useOkapiKy();
@@ -147,128 +132,6 @@ const PiecesPreviewModal = ({
     await generatePieces(submitValues);
   };
 
-  const renderPublicationDate = (piece) => {
-    return <PiecePublicationDate piece={piece} />;
-  };
-
-  /* istanbul ignore next */
-  const formatter = {
-    // If omissionOrigins exist then piece is omitted
-    issueCount: (e) => {
-      return e?.omissionOrigins ? '-' : e.rowIndex + 1;
-    },
-    publicationDate: (e) => renderPublicationDate(e),
-    displaySummary: (e) => {
-      return e?.label;
-    },
-  };
-
-  const renderPiecesTable = () => {
-    return (
-      <Row>
-        <Col xs={12}>
-          <MultiColumnList
-            columnMapping={{
-              issueCount: (
-                <FormattedMessage id="ui-serials-management.ruleset.issueCount" />
-              ),
-              publicationDate: (
-                <FormattedMessage id="ui-serials-management.ruleset.issuePublicationDate" />
-              ),
-              displaySummary: (
-                <FormattedMessage id="ui-serials-management.pieceSets.displaySummary" />
-              ),
-            }}
-            columnWidths={{
-              publicationDate: { min: 100, max: 165 },
-            }}
-            contentData={predictedPieces}
-            formatter={formatter}
-            id="pieces-preview-multi-columns"
-            interactive={false}
-            visibleColumns={['issueCount', 'publicationDate', 'displaySummary']}
-          />
-        </Col>
-      </Row>
-    );
-  };
-
-  const renderEnumerationNumericField = (formatValues, index) => {
-    return (
-      <Row key={`enumeration-numeric-field-${index}`}>
-        <Col xs={2}>
-          <Layout className="textCentered padding-top-gutter">
-            <strong>
-              <FormattedMessage
-                id="ui-serials-management.ruleset.labelIndex"
-                values={{ index: index + 1 }}
-              />
-            </strong>
-          </Layout>
-        </Col>
-        {formatValues?.ruleType?.ruleFormat?.levels?.map((e, i) => {
-          return (
-            <Col key={`${index}-${i}`} xs={2}>
-              <Field
-                component={TextField}
-                id="level-index-label"
-                label={
-                  <FormattedMessage
-                    id="ui-serials-management.ruleset.levelIndex"
-                    values={{ index: i + 1 }}
-                  />
-                }
-                name={`startingValues[${index}].levels[${i}].value`}
-                required
-                type="number"
-                validate={
-                  e?.sequence?.value === 'reset'
-                    ? composeValidators(
-                      requiredValidator,
-                      validateWithinRange(1, e?.units)
-                    )
-                    : requiredValidator
-                }
-              />
-            </Col>
-          );
-        })}
-      </Row>
-    );
-  };
-
-  const renderTemplateStartingValues = () => {
-    return (
-      <div className={css.container}>
-        <Row>
-          <Col xs={12}>
-            <Label>
-              <FormattedMessage id="ui-serials-management.ruleset.valuesToUseForFirstIssue" />
-            </Label>
-          </Col>
-        </Row>
-        <br />
-        {ruleset?.templateConfig?.rules?.map((e, i) => {
-          if (
-            e?.ruleType?.templateMetadataRuleFormat?.value ===
-              'enumeration_numeric' ||
-            e?.ruleType?.templateMetadataRuleFormat === 'enumeration_numeric'
-          ) {
-            // Required so that sonarcloud doesnt flag use of index within key prop
-            const indexCounter = i;
-            return (
-              <div key={`enumeration-numeric-field-container-${indexCounter}`}>
-                {renderEnumerationNumericField(e, i)}
-                <br />
-              </div>
-            );
-          }
-          return null;
-        })}
-      </div>
-    );
-  };
-
   const renderFooter = ({ formState, handleSubmit, handleClose }) => {
     const { invalid, pristine, submitting, values } = formState;
     return (
@@ -316,6 +179,52 @@ const PiecesPreviewModal = ({
     );
   };
 
+  const renderPublicationDate = (piece) => {
+    return <PiecePublicationDate piece={piece} />;
+  };
+
+  /* istanbul ignore next */
+  const formatter = {
+    // If omissionOrigins exist then piece is omitted
+    issueCount: (e) => {
+      return e?.omissionOrigins ? '-' : e.rowIndex + 1;
+    },
+    publicationDate: (e) => renderPublicationDate(e),
+    displaySummary: (e) => {
+      return e?.label;
+    },
+  };
+
+  const renderPiecesTable = () => {
+    return (
+      <Row>
+        <Col xs={12}>
+          <MultiColumnList
+            columnMapping={{
+              issueCount: (
+                <FormattedMessage id="ui-serials-management.ruleset.issueCount" />
+              ),
+              publicationDate: (
+                <FormattedMessage id="ui-serials-management.ruleset.issuePublicationDate" />
+              ),
+              displaySummary: (
+                <FormattedMessage id="ui-serials-management.pieceSets.displaySummary" />
+              ),
+            }}
+            columnWidths={{
+              publicationDate: { min: 100, max: 165 },
+            }}
+            contentData={predictedPieces}
+            formatter={formatter}
+            id="pieces-preview-multi-columns"
+            interactive={false}
+            visibleColumns={['issueCount', 'publicationDate', 'displaySummary']}
+          />
+        </Col>
+      </Row>
+    );
+  };
+
   return (
     <FormModal
       modalProps={{
@@ -329,38 +238,11 @@ const PiecesPreviewModal = ({
       mutators={arrayMutators}
       onSubmit={handleCreation}
     >
-      <Row>
-        <Col xs={4}>
-          <Field
-            backendDateStandard="YYYY-MM-DD"
-            component={Datepicker}
-            id="ruleset-start-date"
-            label={
-              <FormattedMessage id="ui-serials-management.ruleset.startDate" />
-            }
-            name="startDate"
-            required
-            usePortal
-            validate={requiredValidator}
-          />
-        </Col>
-        {allowCreation && (
-          <Col xs={8}>
-            <Field
-              component={TextArea}
-              label={
-                <FormattedMessage id="ui-serials-management.pieceSets.note" />
-              }
-              name="note"
-            />
-          </Col>
-        )}
-      </Row>
-      {!!ruleset?.templateConfig?.rules?.some(
-        (e) => e?.ruleType?.templateMetadataRuleFormat?.value ===
-            'enumeration_numeric' ||
-          e?.ruleType?.templateMetadataRuleFormat === 'enumeration_numeric'
-      ) && renderTemplateStartingValues()}
+      <PiecesPreviewModalForm
+        allowCreation={allowCreation}
+        pieceSets={pieceSets}
+        ruleset={ruleset}
+      />
       {!!predictedPieces && renderPiecesTable()}
     </FormModal>
   );
